backend/requirements.txt
flask
flask-cors
groq
PyMuPDF
openpyxl
bcrypt
PyJWT
gunicorn
